requirements.txt
requests==2.31.0
websocket-client==1.6.4
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0
ccxt==4.1.77
ta-lib==0.4.28
schedule==1.2.0
loguru==0.7.2
pydantic==2.5.2
asyncio==3.4.3
aiohttp==3.9.1
